% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpt05_adj.R
\name{adj_readData}
\alias{adj_readData}
\title{adj}
\usage{
adj_readData(
  file = "data-raw/adj/手工调整数据模板--自然堂事业部1-5月.xlsx",
  conn = tsda::conn_rds("jlrds")
)
}
\arguments{
\item{file}{ile}

\item{conn}{conn}
}
\value{
return
}
\description{
adj
}
\examples{
adj_readData()
}
