% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpt05_adj_ui.R
\name{adj_res_ui_fromDB}
\alias{adj_res_ui_fromDB}
\title{读取adj处理中间表}
\usage{
adj_res_ui_fromDB(
  conn = tsda::conn_rds("jlrds"),
  FBrand = "自然堂",
  FChannel = "商超",
  FYear = 2021,
  FPeriod = 5
)
}
\arguments{
\item{conn}{连接}

\item{FBrand}{品牌}

\item{FChannel}{渠道}

\item{FYear}{年}

\item{FPeriod}{月}
}
\value{
返回
}
\description{
读取adj处理中间表
}
\examples{
adj_res_ui_fromDB()
}
