% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpt_bw2.R
\name{bw2_deal_list}
\alias{bw2_deal_list}
\title{获取链接}
\usage{
bw2_deal_list(conn = tsda::conn_rds("jlrds"), FYear = 2021, FPeriod = 6)
}
\arguments{
\item{conn}{连接}

\item{FYear}{年份}

\item{FPeriod}{月份}
}
\value{
返回值
}
\description{
获取链接
}
\examples{
bw2_deal_list
}
