% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpt_bw2.R
\name{bw2_Filter_F30_Notin}
\alias{bw2_Filter_F30_Notin}
\title{维度筛选30B客户排除}
\usage{
bw2_Filter_F30_Notin(
  conn = tsda::conn_rds("jlrds"),
  FSolutionNumber = "S001",
  FSubNumber = 1
)
}
\arguments{
\item{conn}{连接}

\item{FSolutionNumber}{方案号}

\item{FSubNumber}{子方案号}
}
\value{
返回值
}
\description{
维度筛选30B客户排除
}
\examples{
bw2_Filter_F13_in()
}
