% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpt_bw2.R
\name{bw2_Filter_valueType}
\alias{bw2_Filter_valueType}
\title{维度筛选指标数据,一定是最后使用，不需要and}
\usage{
bw2_Filter_valueType(
  conn = tsda::conn_rds("jlrds"),
  FSolutionNumber = "S001",
  FSubNumber = 1
)
}
\arguments{
\item{conn}{连接}

\item{FSolutionNumber}{方案号}

\item{FSubNumber}{子方案号}
}
\value{
返回值
}
\description{
维度筛选指标数据,一定是最后使用，不需要and
}
\examples{
bw2_Filter_F13_in()
}
