% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpt_bw2.R
\name{bw2_selectData}
\alias{bw2_selectData}
\title{整合查询}
\usage{
bw2_selectData(
  conn = tsda::conn_rds("jlrds"),
  FSolutionNumber = "S001",
  FSubNumber = 1,
  FYear = 2021,
  FPeriod = 6
)
}
\arguments{
\item{conn}{连接}

\item{FSolutionNumber}{方案号}

\item{FSubNumber}{子序号}
}
\value{
返回值
}
\description{
整合查询
}
\examples{
bw2_selectData
}
