% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpt_bw2.R
\name{bw2_sync_data}
\alias{bw2_sync_data}
\title{同步BW2版本的纵表数据}
\usage{
bw2_sync_data(conn = tsda::conn_rds("jlrds"), FYear = 2021, FPeriod = 6)
}
\arguments{
\item{conn}{连接}

\item{FYear}{年}

\item{FPeriod}{月}
}
\value{
返回值
}
\description{
同步BW2版本的纵表数据
}
\examples{

}
