% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly.R
\name{month_DelBal}
\alias{month_DelBal}
\title{删除月报余额}
\usage{
month_DelBal(conn = tsda::conn_rds("jlrds"), Fyear = 2020, Fmonth = 7)
}
\arguments{
\item{conn}{连接}

\item{Fyear}{年份}

\item{Fmonth}{月份}
}
\value{
返回值
}
\description{
删除月报余额
}
\examples{
month_DelBal()
}
