% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly.R
\name{month_DelRpt}
\alias{month_DelRpt}
\title{删除月报数据}
\usage{
month_DelRpt(conn = tsda::conn_rds("jlrds"), Fyear = 2020, Fmonth = 7)
}
\arguments{
\item{conn}{连接}

\item{Fyear}{年份}

\item{Fmonth}{月份}
}
\value{
返回值
}
\description{
删除月报数据
}
\examples{
month_DelRpt()
}
