% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly.R
\name{month_DelStat}
\alias{month_DelStat}
\title{删除月报统计数据}
\usage{
month_DelStat(conn = tsda::conn_rds("jlrds"), Fyear = 2020, Fmonth = 7)
}
\arguments{
\item{conn}{连接}

\item{Fyear}{年份}

\item{Fmonth}{月份}
}
\value{
返回值
}
\description{
删除月报统计数据
}
\examples{
month_DelStat()
}
