% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly.R
\name{month_getBeginBal}
\alias{month_getBeginBal}
\title{获取每个月的期初数}
\usage{
month_getBeginBal(conn = tsda::conn_rds("jlrds"), year = 2020, month = 3)
}
\arguments{
\item{conn}{连接}

\item{year}{年份}

\item{month}{月份}
}
\value{
返回值
}
\description{
获取每个月的期初数
}
\examples{
month_getBeginBal()
}
