% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly.R
\name{month_getDates}
\alias{month_getDates}
\title{获取一周的开始日期与结束日志}
\usage{
month_getDates(conn = tsda::conn_rds("jlrds"), year = 2020, month = 3)
}
\arguments{
\item{conn}{连接}

\item{year}{年份}

\item{month}{月份}
}
\value{
返回值
}
\description{
获取一周的开始日期与结束日志
}
\examples{
month_getDates()
}
