% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly.R
\name{month_getEndBal}
\alias{month_getEndBal}
\title{按月读取余额数}
\usage{
month_getEndBal(conn = tsda::conn_rds("jlrds"), year = 2020, month = 3)
}
\arguments{
\item{conn}{连接}

\item{year}{年份}

\item{month}{月份}
}
\value{
返回值
}
\description{
按月读取余额数
}
\examples{
month_getEndBal()
}
