% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly.R
\name{month_getLastMonth}
\alias{month_getLastMonth}
\title{获取上一年份与月份事情}
\usage{
month_getLastMonth(conn = tsda::conn_rds("jlrds"), year = 2020, month = 3)
}
\arguments{
\item{conn}{连接}

\item{year}{年份}

\item{month}{月份}
}
\value{
返回值
}
\description{
获取上一年份与月份事情
}
\examples{
month_getLastMonth()
}
