% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly.R
\name{month_getRptType}
\alias{month_getRptType}
\title{获取月报的资金类型}
\usage{
month_getRptType(conn = tsda::conn_rds("jlrds"))
}
\arguments{
\item{conn}{连接}
}
\value{
返回
}
\description{
获取月报的资金类型
}
\examples{
month_getRptType()
}
