% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly.R
\name{month_stat}
\alias{month_stat}
\title{处理周报统计信息}
\usage{
month_stat(conn = tsda::conn_rds("jlrds"), year = 2020, month = 3)
}
\arguments{
\item{conn}{连接}

\item{year}{年份}

\item{weekNo}{周号}

\item{type}{类型}
}
\value{
返回值
}
\description{
处理周报统计信息
}
\examples{
week_stat()
}
