% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly.R
\name{month_update}
\alias{month_update}
\title{月报更新}
\usage{
month_update(conn = tsda::conn_rds("jlrds"), Fyear = 2020, Fmonth = 7)
}
\arguments{
\item{conn}{连接}

\item{Fyear}{月份}

\item{Fmonth}{月份}
}
\value{
返回值
}
\description{
月报更新
}
\examples{
month_update()
}
