% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly.R
\name{month_writeMonthBal}
\alias{month_writeMonthBal}
\title{更新月余额数据}
\usage{
month_writeMonthBal(
  conn = tsda::conn_rds("jlrds"),
  year = 2020,
  month = 3,
  value
)
}
\arguments{
\item{conn}{连接}

\item{year}{年份}

\item{month}{月份}

\item{value}{数值}
}
\value{
返回值
}
\description{
更新月余额数据
}
\examples{
month_writeMonthBal()
}
