% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpt31_history_ui.R
\name{mrpt_actualData_read}
\alias{mrpt_actualData_read}
\title{读取历史数据}
\usage{
mrpt_actualData_read(
  conn = tsda::conn_rds("jlrds"),
  FYear = 2020,
  FPeriod = 1,
  FBrand = "美素",
  FChannel = "大客户"
)
}
\arguments{
\item{conn}{连接}

\item{FYear}{年}

\item{FPeriod}{期间}

\item{FBrand}{品牌}

\item{FChannel}{渠道}
}
\value{
返回值
}
\description{
读取历史数据
}
\examples{
mrpt_actualData_read()
}
