% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpt05_adj_ui.R
\name{mrpt_adj_readData_ui}
\alias{mrpt_adj_readData_ui}
\title{读取调整数据}
\usage{
mrpt_adj_readData_ui(conn = tsda::conn_rds("jlrds"), FYear = 2021, FPeriod = 1)
}
\arguments{
\item{conn}{连接}

\item{FYear}{年份}

\item{FPeriod}{期间}
}
\value{
返回值
}
\description{
读取调整数据
}
\examples{
mrpt_adj_readData_ui()
}
