% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpt11_budget.R
\name{mrpt_budget_readData}
\alias{mrpt_budget_readData}
\title{写入成本项目数据}
\usage{
mrpt_budget_readData(
  file = "data-raw/mrpt04/BudgetData/预算整合数据.xlsx",
  conn = tsda::conn_rds("jlrds")
)
}
\arguments{
\item{file}{文件}

\item{conn}{连接}
}
\value{
返回值
}
\description{
写入成本项目数据
}
\examples{
mrpt_budget_readData()
}
