% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpt11_budget.R
\name{mrpt_budget_readData_ByDivision_cumPreCheck}
\alias{mrpt_budget_readData_ByDivision_cumPreCheck}
\title{进行报表数据的以预检验}
\usage{
mrpt_budget_readData_ByDivision_cumPreCheck(
  conn = tsda::conn_rds("jlrds"),
  FBrand = "珀芙研",
  FChannel = "电商",
  FYear = 2021,
  FPeriod = 5,
  FSubChannel = NA
)
}
\arguments{
\item{conn}{连接}

\item{FBrand}{品牌}

\item{FChannel}{渠道}

\item{FYear}{年}

\item{FPeriod}{月份}

\item{FSubChannel}{子渠道}
}
\value{
返回值
}
\description{
进行报表数据的以预检验
}
\examples{
mrpt_budget_readData_ByDivision_cumPreCheck()
}
