% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpt11_budget.R
\name{mrpt_budget_readData_ByDivision_currentPeriod}
\alias{mrpt_budget_readData_ByDivision_currentPeriod}
\title{获取当前期间的预算数}
\usage{
mrpt_budget_readData_ByDivision_currentPeriod(
  file = "data-raw/budget/执行预算_RDS_珀_OK.xlsx",
  conn = tsda::conn_rds("jlrds"),
  sheet = "电商",
  FBrand = "珀芙研",
  FChannel = "电商",
  FYear = 2021,
  FPeriod = 5,
  FSubChannel = NA
)
}
\arguments{
\item{file}{文件}

\item{conn}{连接}

\item{sheet}{页答}

\item{FBrand}{品牌}

\item{FChannel}{渠道}

\item{FYear}{年}

\item{FPeriod}{月}

\item{FSubChannel}{子渠道}
}
\value{
返回值
}
\description{
获取当前期间的预算数
}
\examples{
mrpt_budget_readData_ByDivision_currentPeriod()
}
