% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpt11_budget_ui.R
\name{mrpt_budget_readFromDB_ByBrandChannel}
\alias{mrpt_budget_readFromDB_ByBrandChannel}
\title{读取预算数据}
\usage{
mrpt_budget_readFromDB_ByBrandChannel(
  conn = tsda::conn_rds("jlrds"),
  FBrand = "珀芙研",
  FChannel = "药房",
  FYear = 2021,
  FPeriod = 5
)
}
\arguments{
\item{conn}{连接}

\item{FBrand}{品牌}

\item{FChannel}{渠道}

\item{FYear}{年}

\item{FPeriod}{月}
}
\value{
返回值
}
\description{
读取预算数据
}
\examples{
mrpt_budget_readFromDB_ByBrandChannel()
}
