% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpt01_bw.R
\name{mrpt_bw_b001}
\alias{mrpt_bw_b001}
\title{BW001报表的处理}
\usage{
mrpt_bw_b001(file = "data-raw/mrpt04/mrpt01_bw/B001_04_data.xlsx")
}
\arguments{
\item{file}{文件名}
}
\value{
返回值
}
\description{
BW001报表的处理
}
\examples{
mrpt_bw_b001()
}
