% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpt01_bw.R
\name{mrpt_bw_b002}
\alias{mrpt_bw_b002}
\title{BW001报表的处理}
\usage{
mrpt_bw_b002(file = "data-raw/mrpt04/mrpt01_bw/B002_04_data.xlsx", skip = 1)
}
\arguments{
\item{file}{文件名}

\item{skip}{忽略第一行}
}
\value{
返回值
}
\description{
BW001报表的处理
}
\examples{
mrpt_bw_b002()
}
