% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpt01_bw.R
\name{mrpt_bw_deal}
\alias{mrpt_bw_deal}
\title{BW报表处理规则}
\usage{
mrpt_bw_deal(
  file = "data-raw/mrpt04/BWdata/B001_04.xlsx",
  FYear = 2021,
  FPeriod = 4,
  conn = tsda::conn_rds("jlrds"),
  skip = 1
)
}
\arguments{
\item{file}{文件名}

\item{FYear}{年份}

\item{FPeriod}{月份}

\item{conn}{连接}

\item{skip}{忽略行}
}
\value{
返回值
}
\description{
BW报表处理规则
}
\examples{
mrpt_bw_deal()
}
