% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpt01_bw.R
\name{mrpt_bw_deal_batch}
\alias{mrpt_bw_deal_batch}
\title{针对数据进行批量处理}
\usage{
mrpt_bw_deal_batch(
  files = "data-raw/mrpt04/BWdata",
  FYear = 2021,
  FPeriod = 4,
  conn = tsda::conn_rds("jlrds"),
  skip = 1
)
}
\arguments{
\item{files}{文件名}

\item{FYear}{年份}

\item{FPeriod}{月份}

\item{conn}{连接}

\item{skip}{跳出行}
}
\value{
返回值
}
\description{
针对数据进行批量处理
}
\examples{
mrpt_bw_deal_batch()
}
