% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpt01_bw.R
\name{mrpt_bw_dim_bySolutionNumber_checked}
\alias{mrpt_bw_dim_bySolutionNumber_checked}
\title{返回BW报表的维度信息}
\usage{
mrpt_bw_dim_bySolutionNumber_checked(
  conn = tsda::conn_rds("jlrds"),
  solutionNumber = "B001"
)
}
\arguments{
\item{conn}{连接}

\item{solutionNumber}{方案号}
}
\value{
返回值
}
\description{
返回BW报表的维度信息
}
\examples{
mrpt_bw_dim_bySolutionNumber_checked
}
