% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpt01_bw_ui.R
\name{mrpt_bw_ds_data}
\alias{mrpt_bw_ds_data}
\title{获取BW数据源}
\usage{
mrpt_bw_ds_data(conn = tsda::conn_rds("jlrds"), FYear = 2021, FPeriod = 5)
}
\arguments{
\item{conn}{连接}

\item{FYear}{年份}

\item{FPeriod}{期间}
}
\value{
返回值
}
\description{
获取BW数据源
}
\examples{
mrpt_bw_ds_data()
}
