% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpt01_bw.R
\name{mrpt_bw_getSolutionByFile}
\alias{mrpt_bw_getSolutionByFile}
\title{根据文件获取方案号}
\usage{
mrpt_bw_getSolutionByFile(
  file_name = "data-raw/mrpt04/mrpt01_bw/B001_04_data.xlsx"
)
}
\arguments{
\item{file_name}{文件名}
}
\value{
返回值
}
\description{
根据文件获取方案号
}
\examples{
mrpt_bw_getSolutionByFile()
}
