% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpt01_bw.R
\name{mrpt_bw_rule_brandChannel}
\alias{mrpt_bw_rule_brandChannel}
\title{获取规则表的品牌及渠道}
\usage{
mrpt_bw_rule_brandChannel(
  conn = tsda::conn_rds("jlrds"),
  solutionNumber = "B001"
)
}
\arguments{
\item{conn}{连接}

\item{solutionNumber}{方案号}
}
\value{
返回值
}
\description{
获取规则表的品牌及渠道
}
\examples{
mrpt_bw_rule_brandChannel()
}
