% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpt01_bw_ui.R
\name{mrpt_bw_ui_businessRule}
\alias{mrpt_bw_ui_businessRule}
\title{BW报表的业务处理规则}
\usage{
mrpt_bw_ui_businessRule(conn = tsda::conn_rds("jlrds"))
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
BW报表的业务处理规则
}
\examples{
mrpt_bw_ui_businessRule()
}
