% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpt01_bw_ui.R
\name{mrpt_bw_ui_getDimName}
\alias{mrpt_bw_ui_getDimName}
\title{BW报表的维度名称}
\usage{
mrpt_bw_ui_getDimName(conn = tsda::conn_rds("jlrds"))
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
BW报表的维度名称
}
\examples{
mrpt_bw_ui_getDimName()
}
