% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpt01_bw_ui.R
\name{mrpt_bw_ui_getHeadingName}
\alias{mrpt_bw_ui_getHeadingName}
\title{查看BW报表表头}
\usage{
mrpt_bw_ui_getHeadingName(conn = tsda::conn_rds("jlrds"))
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
查看BW报表表头
}
\examples{
mrpt_bw_ui_getHeadingName()
}
