% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpt01_bw.R
\name{mrpt_bw_value}
\alias{mrpt_bw_value}
\title{返回SAP报表的固定表头}
\usage{
mrpt_bw_value(conn = tsda::conn_rds("jlrds"))
}
\arguments{
\item{conn}{连接信息从数据库读取}
}
\value{
返回值
}
\description{
返回SAP报表的固定表头
}
\examples{
mrpt_bw_value()
}
