% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpt01_bw.R
\name{mrpt_bw_value_bySolutionNumber_checked}
\alias{mrpt_bw_value_bySolutionNumber_checked}
\title{获取BW报表表头}
\usage{
mrpt_bw_value_bySolutionNumber_checked(
  conn = tsda::conn_rds("jlrds"),
  solutionNumber = "B001"
)
}
\arguments{
\item{conn}{连接}

\item{solutionNumber}{方案号}
}
\value{
返回值
}
\description{
获取BW报表表头
}
\examples{
mrpt_bw_value_bySolutionNumber_checked()
}
