% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpt01_bw.R
\name{mrpt_bw_value_bySolutionNumber_unChecked}
\alias{mrpt_bw_value_bySolutionNumber_unChecked}
\title{获取表头信息}
\usage{
mrpt_bw_value_bySolutionNumber_unChecked(
  conn = tsda::conn_rds("jlrds"),
  solutionNumber = "B001"
)
}
\arguments{
\item{conn}{连接}

\item{solutionNumber}{方案号}
}
\value{
返回值
}
\description{
获取表头信息
}
\examples{
mrpt_bw_heading_bySolutionNumber_unChecked()
}
