% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpt02_sap_ui.R
\name{mrpt_ds_ui_sapData}
\alias{mrpt_ds_ui_sapData}
\title{读取SAP数据源信息}
\usage{
mrpt_ds_ui_sapData(conn = tsda::conn_rds("jlrds"), FYear = 2021, FPeriod = 5)
}
\arguments{
\item{conn}{连接}

\item{FYear}{年份}

\item{FPeriod}{月份}
}
\value{
返回值
}
\description{
读取SAP数据源信息
}
\examples{
mrpt_ds_ui_sapData()
}
