% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpt01_bw.R
\name{mrpt_getValue_selected}
\alias{mrpt_getValue_selected}
\title{获取方案对应的指标变更}
\usage{
mrpt_getValue_selected(conn = tsda::conn_rds("jlrds"), solutionNumber = "B001")
}
\arguments{
\item{conn}{连接}

\item{solutionNumber}{方案号}
}
\value{
返回值
}
\description{
获取方案对应的指标变更
}
\examples{
mrpt_getValue_selected()
}
