% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpt31_history.R
\name{mrpt_HistoryData_ReadAll}
\alias{mrpt_HistoryData_ReadAll}
\title{读取历史数据}
\usage{
mrpt_HistoryData_ReadAll(
  conn = tsda::conn_rds("jlrds"),
  toc_name = "data-raw/hist/hist_toc.xlsx"
)
}
\arguments{
\item{conn}{连接}

\item{toc_name}{目录文件名}
}
\value{
返回值
}
\description{
读取历史数据
}
\examples{
mrpt_HistoryData_ReadAll()
}
