% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpt31_history.R
\name{mrpt_HistoryData_ReadOne}
\alias{mrpt_HistoryData_ReadOne}
\title{读取历史数据真实数据}
\usage{
mrpt_HistoryData_ReadOne(
  conn = tsda::conn_rds("jlrds"),
 
    file_name = "data-raw/hist/01-美素大客户事业部2019历史管报数据_reviewed.xlsx",
  sheet_name = "美素美妆",
  FBrand = "美素",
  FChannel = "美妆",
  FYear = 2019,
  FSubChannel = NA
)
}
\arguments{
\item{conn}{连接}

\item{file_name}{文件名}

\item{sheet_name}{页签名}

\item{FBrand}{品牌}

\item{FChannel}{渠道}

\item{FYear}{年份}

\item{FSubChannel}{子渠道}
}
\value{
返回值
}
\description{
读取历史数据真实数据
}
\examples{
mrpt_HistoryData_ReadOne()
}
