% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpt_md_budget.R
\name{mrpt_md_brand}
\alias{mrpt_md_brand}
\title{获取品牌信息}
\usage{
mrpt_md_brand(conn = tsda::conn_rds("jlrds"))
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
获取品牌信息
}
\examples{
mrpt_md_brand()
}
