% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpt_md_budget.R
\name{mrpt_md_channel}
\alias{mrpt_md_channel}
\title{获取渠道信息}
\usage{
mrpt_md_channel(conn = tsda::conn_rds("jlrds"))
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
获取渠道信息
}
\examples{
mrpt_md_channel()
}
