% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpt02_sap_ui.R
\name{mrpt_md_ui_costCenter}
\alias{mrpt_md_ui_costCenter}
\title{找到成本中心及渠道费用分配标准}
\usage{
mrpt_md_ui_costCenter(
  conn = tsda::conn_rds("jlrds"),
  FYear = 2021,
  FPeriod = 5
)
}
\arguments{
\item{conn}{连接}

\item{FYear}{年份}

\item{FPeriod}{月}
}
\value{
返回值
}
\description{
找到成本中心及渠道费用分配标准
}
\examples{
mrpt_md_ui_costCenter()
}
