% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpt02_sap_ui.R
\name{mrpt_md_ui_division}
\alias{mrpt_md_ui_division}
\title{事业部定义表}
\usage{
mrpt_md_ui_division(conn = tsda::conn_rds("jlrds"))
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
事业部定义表
}
\examples{
mrpt_md_ui_division()
}
