% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpt02_sap_ui.R
\name{mrpt_md_ui_rptItem}
\alias{mrpt_md_ui_rptItem}
\title{报表项目表}
\usage{
mrpt_md_ui_rptItem(conn = tsda::conn_rds("jlrds"))
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
报表项目表
}
\examples{
mrpt_md_ui_rptItem()
}
