% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpt31_history.R
\name{mrpt_rptItem_getNumber}
\alias{mrpt_rptItem_getNumber}
\title{读取报表项目}
\usage{
mrpt_rptItem_getNumber(conn = tsda::conn_rds("jlrds"))
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
读取报表项目
}
\examples{
mrpt_rptItem_getNumber()
}
