% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpt21_discount.R
\name{mrpt_rule_discount}
\alias{mrpt_rule_discount}
\title{读取管报的供货折扣，用于计算公司零售额}
\usage{
mrpt_rule_discount(conn = tsda::conn_rds("jlrds"), FBrand = "珀芙研")
}
\arguments{
\item{conn}{连接}

\item{FBrand}{品牌}
}
\value{
返回值
}
\description{
读取管报的供货折扣，用于计算公司零售额
}
\examples{
mrpt_rule_discount()
}
