% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpt02_sap_ui.R
\name{mrpt_ui_itemMapping_costRpt}
\alias{mrpt_ui_itemMapping_costRpt}
\title{成本项目及报表项目对照表}
\usage{
mrpt_ui_itemMapping_costRpt(
  conn = tsda::conn_rds("jlrds"),
  FYear = 2021,
  FPeriod = 5
)
}
\arguments{
\item{conn}{连接}

\item{FYear}{年份}

\item{FPeriod}{月份}
}
\value{
返回值
}
\description{
成本项目及报表项目对照表
}
\examples{
mrpt_ui_itemMapping_costRpt()
}
