% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/own.R
\name{own_deal}
\alias{own_deal}
\title{按固定格式处理数据}
\usage{
own_deal(
  conn = tsda::conn_rds("jlrds"),
  FYear = 2020,
  FMonth = 7,
  FUnit = "yuan",
  digit = 2
)
}
\arguments{
\item{conn}{连接}

\item{FYear}{年份}

\item{FMonth}{月份}

\item{FUnit}{金额单位}

\item{digit}{保留小数}
}
\value{
返回值
}
\description{
按固定格式处理数据
}
\examples{
own_deal()
}
