% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/own.R
\name{own_getBcsRpt}
\alias{own_getBcsRpt}
\title{获取合并报表数据}
\usage{
own_getBcsRpt(
  conn = tsda::conn_rds("jlrds"),
  FYear = 2021,
  FMonth = 3,
  digit = 2
)
}
\arguments{
\item{conn}{连接}

\item{FYear}{年份}

\item{FMonth}{月份}

\item{digit}{小数位数}

\item{FUnit}{单位}
}
\value{
返回值
}
\description{
获取合并报表数据
}
\examples{
own_getBcsRpt()
}
