% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/own.R
\name{own_getRptTpl}
\alias{own_getRptTpl}
\title{查询自有资金模板表}
\usage{
own_getRptTpl(conn = tsda::conn_rds("jlrds"))
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
查询自有资金模板表
}
\examples{
own_getRptTpl()
}
