% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily.R
\name{rpt_daily_readExcel}
\alias{rpt_daily_readExcel}
\title{读取日报数据}
\usage{
rpt_daily_readExcel(file = "data-raw/jala_rpt.xlsx", sheet = "daily")
}
\arguments{
\item{file}{文件}

\item{sheet}{页答}
}
\value{
返回值
}
\description{
读取日报数据
}
\examples{
rpt_daily_readExcel()
}
