% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily.R
\name{rpt_daily_selectDb}
\alias{rpt_daily_selectDb}
\title{查询日志数据}
\usage{
rpt_daily_selectDb(
  conn = tsda::conn_rds("jlrds"),
  FStartDate = "2020-06-01",
  FEndDate = "2020-06-30",
  FLevel = 0,
  format = "rpt"
)
}
\arguments{
\item{conn}{连接}

\item{FStartDate}{开始日期}

\item{FEndDate}{结束日期}

\item{FLevel}{级次}

\item{format}{格式}
}
\value{
返回值
}
\description{
查询日志数据
}
\examples{
rpt_daily_selectDb()
}
