% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily.R
\name{rpt_daily_writeDb}
\alias{rpt_daily_writeDb}
\title{资金日报写入数据库}
\usage{
rpt_daily_writeDb(
  file = "data-raw/jala_rpt.xlsx",
  sheet = "daily",
  conn = tsda::conn_rds("jlrds")
)
}
\arguments{
\item{file}{文件}

\item{sheet}{页答}

\item{conn}{连接信息}
}
\value{
返回值
}
\description{
资金日报写入数据库
}
\examples{
rpt_daily_writeDb()
}
