% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{rpt_percent}
\alias{rpt_percent}
\title{计算数值百分比}
\usage{
rpt_percent(value, base, digit = 2)
}
\arguments{
\item{value}{数据}

\item{base}{基数}

\item{digit}{小数位数}
}
\value{
返回值
}
\description{
计算数值百分比
}
\examples{
rpt_percent()
}
