% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpt02_sap.R
\name{sap_costCenter_standardize}
\alias{sap_costCenter_standardize}
\title{针对SAP成本中心进行标准化处理}
\usage{
sap_costCenter_standardize(file = "data-raw/rpt_mngr/成本中心划分总表.xlsx")
}
\arguments{
\item{file}{文件名}
}
\value{
返回值
}
\description{
针对SAP成本中心进行标准化处理
}
\examples{
sap_costCenter_standardize()
}
