% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpt02_sap.R
\name{sap_getBrandChannel_byName}
\alias{sap_getBrandChannel_byName}
\title{根据渠道类型返回名称}
\usage{
sap_getBrandChannel_byName(
  conn = tsda::conn_rds("jlrds"),
  FChannelName = "COMO市场"
)
}
\arguments{
\item{conn}{连接}

\item{FChannelName}{渠道描述}
}
\value{
返回值
}
\description{
根据渠道类型返回名称
}
\examples{
sap_getBrandChannel_byName()
}
