% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpt02_sap_ui.R
\name{sap_res_ui_fromDB}
\alias{sap_res_ui_fromDB}
\title{SAP处理结果表}
\usage{
sap_res_ui_fromDB(
  conn = tsda::conn_rds("jlrds"),
  FBrand = "自然堂",
  FChannel = "商超",
  FYear = 2021,
  FPeriod = 5
)
}
\arguments{
\item{conn}{连接}

\item{FBrand}{品牌}

\item{FChannel}{渠道}

\item{FYear}{年份}

\item{FPeriod}{月份}
}
\value{
返回值
}
\description{
SAP处理结果表
}
\examples{
sap_res_ui_fromDB()
}
