% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpt02_sap.R
\name{sap_rpt_costItem_readData}
\alias{sap_rpt_costItem_readData}
\title{写入成本项目数据}
\usage{
sap_rpt_costItem_readData(
  file = "data-raw/rpt_mngr/SAP成本要素对应关系表.xlsx",
  FYear = 2021,
  FPeriod = 4,
  conn = tsda::conn_rds("jlrds")
)
}
\arguments{
\item{file}{文件}

\item{FYear}{年}

\item{FPeriod}{月}

\item{conn}{连接}
}
\value{
返回值
}
\description{
写入成本项目数据
}
\examples{
sap_rpt_costItem_readData()
}
