% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpt02_sap.R
\name{sap_rpt_readData}
\alias{sap_rpt_readData}
\title{SAP报表读取数据}
\usage{
sap_rpt_readData(
  file = "data-raw/mrpt04/SAPdata/SAP_04_2021V2.XLSX",
  FYear = 2021,
  FPeriod = 4,
  conn = tsda::conn_rds("jlrds")
)
}
\arguments{
\item{file}{文件}

\item{FYear}{年}

\item{FPeriod}{月}

\item{conn}{连接}
}
\value{
返回值
}
\description{
SAP报表读取数据
}
\examples{
sap_rpt_readData()
}
