% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekly.R
\name{week_deal}
\alias{week_deal}
\title{按周处理数据}
\usage{
week_deal(
  conn = tsda::conn_rds("jlrds"),
  year = 2020,
  weekNo = 9,
  type = "jala"
)
}
\arguments{
\item{conn}{连接}

\item{year}{年份}

\item{weekNo}{周号}

\item{type}{类型}
}
\value{
返回值
}
\description{
按周处理数据
}
\examples{
week_deal()
}
