% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekly.R
\name{week_DelBal}
\alias{week_DelBal}
\title{针对周的余额数据执行删除处理}
\usage{
week_DelBal(
  conn = tsda::conn_rds("jlrds"),
  Fyear = 2020,
  FweekNo = 30,
  Ftype = "nature"
)
}
\arguments{
\item{conn}{连接}

\item{Fyear}{年份}

\item{FweekNo}{周次}

\item{Ftype}{类型}
}
\value{
返回值
}
\description{
针对周的余额数据执行删除处理
}
\examples{
week_DelBal()
}
