% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekly.R
\name{week_DelRpt}
\alias{week_DelRpt}
\title{删除周报数据}
\usage{
week_DelRpt(
  conn = tsda::conn_rds("jlrds"),
  Fyear = 2020,
  FweekNo = 30,
  Ftype = "nature"
)
}
\arguments{
\item{conn}{连接}

\item{Fyear}{年份}

\item{FweekNo}{周次}

\item{Ftype}{类型}
}
\value{
返回值
}
\description{
删除周报数据
}
\examples{
week_DelRpt()
}
