% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekly.R
\name{week_DelStat}
\alias{week_DelStat}
\title{删除统计信息}
\usage{
week_DelStat(
  conn = tsda::conn_rds("jlrds"),
  Fyear = 2020,
  FweekNo = 30,
  Ftype = "nature"
)
}
\arguments{
\item{conn}{连接}

\item{Fyear}{年份}

\item{FweekNo}{周次}

\item{Ftype}{类型}
}
\value{
返回值
}
\description{
删除统计信息
}
\examples{
week_DelStat()
}
