% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekly.R
\name{week_getDateList}
\alias{week_getDateList}
\title{获取周号信息}
\usage{
week_getDateList(conn = tsda::conn_rds("jlrds"), year = 2020, Ftype = "jala")
}
\arguments{
\item{conn}{连接}

\item{year}{年份}

\item{Ftype}{类型}
}
\value{
返回值
}
\description{
获取周号信息
}
\examples{
week_getDateList()
}
