% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekly.R
\name{week_getDates}
\alias{week_getDates}
\title{获取一周的开始日期与结束日志}
\usage{
week_getDates(
  conn = tsda::conn_rds("jlrds"),
  year = 2020,
  weekNo = 9,
  type = "jala"
)
}
\arguments{
\item{conn}{连接}

\item{year}{年份}

\item{weekNo}{周号}

\item{type}{类型}
}
\value{
返回值
}
\description{
获取一周的开始日期与结束日志
}
\examples{
week_getDates()
}
