% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekly.R
\name{week_getEndBal}
\alias{week_getEndBal}
\title{按周读取余额数}
\usage{
week_getEndBal(
  conn = tsda::conn_rds("jlrds"),
  year = 2020,
  weekNo = 8,
  type = "jala"
)
}
\arguments{
\item{conn}{连接}

\item{year}{年份}

\item{weekNo}{周号}

\item{type}{类型}
}
\value{
返回值
}
\description{
按周读取余额数
}
\examples{
week_getEndBal()
}
