% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekly.R
\name{week_getLastWeek}
\alias{week_getLastWeek}
\title{获取上一周年份与周事情}
\usage{
week_getLastWeek(
  conn = tsda::conn_rds("jlrds"),
  year = 2020,
  weekNo = 9,
  type = "jala"
)
}
\arguments{
\item{conn}{连接}

\item{year}{年份}

\item{weekNo}{周号}

\item{type}{类型}
}
\value{
返回值
}
\description{
获取上一周年份与周事情
}
\examples{
week_getLastWeek()
}
