% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekly.R
\name{week_getRptType}
\alias{week_getRptType}
\title{返回周报的报表类型}
\usage{
week_getRptType(conn = tsda::conn_rds("jlrds"))
}
\arguments{
\item{conn}{连接}
}
\value{
返回
}
\description{
返回周报的报表类型
}
\examples{
week_getRptType()
}
