% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekly.R
\name{week_selectDb}
\alias{week_selectDb}
\title{从数据库中读取周报数据}
\usage{
week_selectDb(
  conn = tsda::conn_rds("jlrds"),
  year = 2020,
  weekNo = 10,
  type = "jala"
)
}
\arguments{
\item{conn}{连接}

\item{year}{年份}

\item{weekNo}{周号}

\item{type}{类型}
}
\value{
返回值
}
\description{
从数据库中读取周报数据
}
\examples{
week_selectDb()
}
