% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekly.R
\name{week_stat}
\alias{week_stat}
\title{处理周报统计信息}
\usage{
week_stat(
  conn = tsda::conn_rds("jlrds"),
  year = 2020,
  weekNo = 10,
  type = "jala"
)
}
\arguments{
\item{conn}{连接}

\item{year}{年份}

\item{weekNo}{周号}

\item{type}{类型}
}
\value{
返回值
}
\description{
处理周报统计信息
}
\examples{
week_stat()
}
