% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekly.R
\name{week_update}
\alias{week_update}
\title{完善后的周报更新逻辑}
\usage{
week_update(
  conn = tsda::conn_rds("jlrds"),
  Fyear = 2020,
  FweekNo = 30,
  Ftype = "nature"
)
}
\arguments{
\item{conn}{连接}

\item{Fyear}{周年份}

\item{FweekNo}{周次}

\item{Ftype}{类型}
}
\value{
返回值
}
\description{
完善后的周报更新逻辑
}
\examples{
week_update()
}
