% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekly.R
\name{week_writeWeekBal}
\alias{week_writeWeekBal}
\title{更新周余额数据}
\usage{
week_writeWeekBal(
  conn = tsda::conn_rds("jlrds"),
  year = 2020,
  weekNo = 9,
  value,
  type = "jala"
)
}
\arguments{
\item{conn}{连接}

\item{year}{年份}

\item{weekNo}{周号}

\item{value}{数值}

\item{type}{类型}
}
\value{
返回值
}
\description{
更新周余额数据
}
\examples{
week_writeWeekBal()
}
