% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekly.R
\name{weekRpt_selectDB}
\alias{weekRpt_selectDB}
\title{查询周报数据}
\usage{
weekRpt_selectDB(
  conn = tsda::conn_rds("jlrds"),
  year = 2020,
  startWeekNo = 9,
  endWeekNo = 12,
  AmtType = c("本周发生额", "较上周变动额", "较上周变动\%"),
  FLevel = 0,
  FType = "jala",
  Amt_Unit = "wan"
)
}
\arguments{
\item{conn}{连接}

\item{year}{年份}

\item{startWeekNo}{开始周号}

\item{endWeekNo}{结束周号}

\item{AmtType}{字段类型向量}

\item{FLevel}{级次}

\item{FType}{类型}

\item{Amt_Unit}{金额单位}
}
\value{
返回值
}
\description{
查询周报数据
}
\examples{
weekRpt_selectDB()
}
