% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly.R
\name{monthRpt_selectDB}
\alias{monthRpt_selectDB}
\title{读取月报数据}
\usage{
monthRpt_selectDB(
  conn = tsda::conn_rds("jlrds"),
  year = 2020,
  AmtType = c("本月发生额", "较上月变动额", "较上月变动\%"),
  FLevel = 0,
  amtUnit = "wan"
)
}
\arguments{
\item{conn}{连接}

\item{year}{年份}

\item{AmtType}{金额类型}

\item{FLevel}{级次}

\item{amtUnit}{金额单位默认万元}
}
\value{
返回值
}
\description{
读取月报数据
}
\examples{
monthRpt_selectDB()
}
